Added exception checking and rollback of transactions.

+++ inventory/src/main/java/com/netcracker/metsko/dao/implementation/GenericDaoImpl.java
@@ -17,8 +17,6 @@ public class GenericDaoImpl<T, Long extends Serializable> implements GenericDao<
 
     private Class<T> tClass;
 
-
-
     public GenericDaoImpl() {
         ParameterizedType genericSuperclass = (ParameterizedType) getClass()
                 .getGenericSuperclass();
@@ -27,21 +25,55 @@ public class GenericDaoImpl<T, Long extends Serializable> implements GenericDao<
 
     @Override
     public void create(T newObject) throws SQLException {
-        tx.begin();
-        entityManager.persist(newObject);
-        tx.commit();
+        try {
+            if (entityManager.contains(newObject)) {
+                System.out.println("The object exists in DB\n");
+            } else {
+                tx.begin();
+                entityManager.persist(newObject);
+                tx.commit();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
 
+            if (tx != null) {
+                try {
+
+                    System.err.print("Transaction is being rolled back");
+                    tx.rollback();
+                } catch (Exception exception) {
+
+                    exception.printStackTrace();
+                }
+            }
+        }
     }
-
-
 
     @Override
     public T update(final T objectToUpdate) throws SQLException {
-        tx.begin();
-        T t=entityManager.merge(objectToUpdate);
-        tx.commit();
-        return t;
-
+        try {
+            if(entityManager.contains(objectToUpdate)) {
+                tx.begin();
+                T t = entityManager.merge(objectToUpdate);
+                tx.commit();
+                return t;
+            }
+            else {
+                System.err.println("There is no such object in DB\n");
+            }
+        }catch (Exception e)
+        {
+            e.printStackTrace();
+            if (tx != null){
+                try {
+                    System.err.print("Transaction is being rolled back\n");
+                    tx.rollback();
+                } catch (Exception exception) {
+                    exception.printStackTrace();
+                }
+            }
+        }
+        return objectToUpdate;
     }
 
     @Override
@@ -51,9 +83,27 @@ public class GenericDaoImpl<T, Long extends Serializable> implements GenericDao<
 
     @Override
     public void delete(T objectToDelete) throws SQLException {
-        tx.begin();
-        entityManager.remove(objectToDelete);
-        tx.commit();
+        try {
+            if(entityManager.contains(objectToDelete)) {
+
+                tx.begin();
+                entityManager.remove(objectToDelete);
+                tx.commit();
+            }
+            else {
+                System.err.println("There is no such object in DB\n");
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            if (tx != null) {
+                try {
+                    System.err.println("Transaction is being rolled back\n");
+                    tx.rollback();
+                } catch (Exception exception) {
+                    exception.printStackTrace();
+                }
+            }
+        }
     }
 
     public EntityManager getEntityManager() {
